Move FileList class to Utils directory
